fix: Resize navigation icons and fonts
This commit makes every icon and text in the navigation section to be smaller and more in line with the original website.

Modified the menu sidenav's width from `w-[28rem]` to `w-[22rem]` to fit within mobile screens.

Also has the currency dropdown to have the triangle icon beside it on mobile screens.